chore(utxo-lib): move payments related types
