+++ packages/utxo-lib/src/payments/p2pk.ts
@@ -5,7 +5,7 @@ import * as typef from 'typeforce';
 import { bitcoin as BITCOIN_NETWORK } from '../networks';
 import * as bscript from '../script';
 import * as lazy from './lazy';
-import type { Payment, PaymentOpts, StackFunction } from './index';
+import { Payment, PaymentOpts, StackFunction } from '../types';
 
 const { OPS } = bscript;
