+++ packages/utxo-lib/src/payments/p2wsh.ts
@@ -7,7 +7,7 @@ import * as bcrypto from '../crypto';
 import { bitcoin as BITCOIN_NETWORK } from '../networks';
 import * as bscript from '../script';
 import * as lazy from './lazy';
-import type { Payment, PaymentOpts, StackElement, StackFunction } from './index';
+import { Payment, PaymentOpts, StackElement, StackFunction } from '../types';
 
 const { OPS } = bscript;
 
@@ -73,7 +73,7 @@ export function p2wsh(a: Payment, opts?: PaymentOpts): Payment {
         network = (a.redeem && a.redeem.network) || BITCOIN_NETWORK;
     }
 
-    const o: Payment = { network };
+    const o: Payment = { name: 'p2wsh', network };
 
     lazy.prop(o, 'address', () => {
         if (!o.hash) return;
@@ -132,7 +132,7 @@ export function p2wsh(a: Payment, opts?: PaymentOpts): Payment {
 
     // extended validation
     if (opts.validate) {
-        let hash: Buffer = Buffer.from([]);
+        let hash = Buffer.from([]);
         if (a.address) {
             const { prefix, version, data } = _address();
             if (prefix !== network.bech32)
