+++ packages/utxo-lib/src/payments/sstxchange.ts
@@ -3,7 +3,7 @@ import * as bs58check from '../bs58check';
 import { decred as DECRED_NETWORK } from '../networks';
 import * as bscript from '../script';
 import * as lazy from './lazy';
-import type { Payment, PaymentOpts, Stack } from './index';
+import { Payment, PaymentOpts, Stack } from '../types';
 
 const { OPS } = bscript;
 
@@ -28,7 +28,7 @@ export function sstxchange(a: Payment, opts?: PaymentOpts): Payment {
     const _address = lazy.value(() => bs58check.decodeAddress(a.address!, a.network));
 
     const network = a.network || DECRED_NETWORK;
-    const o = { name: 'sstxchange', network } as Payment;
+    const o: Payment = { name: 'sstxchange', network };
 
     lazy.prop(o, 'address', () => {
         if (!o.hash) return;
