+++ packages/utxo-lib/src/payments/sstxsh.ts
@@ -3,7 +3,7 @@ import * as bs58check from '../bs58check';
 import { decred as DECRED_NETWORK } from '../networks';
 import * as bscript from '../script';
 import * as lazy from './lazy';
-import type { Payment, PaymentOpts } from './index';
+import { Payment, PaymentOpts } from '../types';
 
 const { OPS } = bscript;
 
@@ -26,7 +26,7 @@ export function sstxsh(a: Payment, opts?: PaymentOpts): Payment {
     );
 
     const network = a.network || DECRED_NETWORK;
-    const o = { name: 'sstxsh', network } as Payment;
+    const o: Payment = { name: 'sstxsh', network };
     const _address = lazy.value(() => bs58check.decodeAddress(a.address!, network));
 
     lazy.prop(o, 'address', () => {
@@ -44,7 +44,7 @@ export function sstxsh(a: Payment, opts?: PaymentOpts): Payment {
 
     // extended validation
     if (opts.validate) {
-        let hash: Buffer = Buffer.from([]);
+        let hash = Buffer.from([]);
         if (a.address) {
             const { version, hash: aHash } = _address();
             if (version !== network.scriptHash)
