+++ packages/utxo-lib/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from './payments';
+export * from './typeforce';
