+++ packages/utxo-lib/tests/payments.test.ts
@@ -1,5 +1,5 @@
 import * as PAYMENTS from '../src/payments';
-import type { PaymentCreator } from '../src/payments';
+import type { PaymentCreator } from '../src/types';
 import * as utils from './payments.utils';
 
 (
